401 aborts
instead of redirecting to `return view('dashboard');` when the viewer is not an admin, instead use `abort(401);`. this shouldnt actually matter, as the auth middleware should handle this, but it's more correct for edge cases.

ALSO: there was a typo in the @endauth directive in `../views/users/index.blade.php` which was fixed