JS user password generation
perfectly adequate solution to address #17 , populates the password fields on the admin user creation page with cryptographically random passwords.

i will revert this is pushed further to handle this server-side. i *really* dont think we should do that when this is so much easier, introduces no dependencies, and works just as well.